--- docs/api/renewal.rst
@@ -1,5 +1,0 @@
-:mod:`certbot.renewal`
---------------------------
-
-.. automodule:: certbot.renewal
-   :members:
